+++ actinia_module_plugin/core/templates/global_templates.py
@@ -32,10 +32,16 @@ from actinia_module_plugin.resources.templating import pcTplEnv
 
 
 def getAll():
+    """This method creates a list of all global templates stored as file
+    to return the list to the api method listing all templates.
+    """
     tpl_list = pcTplEnv.list_templates(filter_func=filter_func)
     tpl_list = [i.replace('.json', '').split('/')[-1] for i in tpl_list]
     return tpl_list
 
 
 def getTemplate(template_id):
+    """This method uses other core methods to return a rendered template stored
+    as file to return it to the api method showing a certain template.
+    """
     return render_template(template_id)
